+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/util/SentryPluginUtilsTest.kt
@@ -3,12 +3,20 @@ package io.sentry.android.gradle.util
 import io.sentry.android.gradle.testutil.createTestAndroidProject
 import io.sentry.android.gradle.testutil.createTestProguardProject
 import io.sentry.android.gradle.util.SentryPluginUtils.capitalizeUS
+import io.sentry.android.gradle.util.SentryPluginUtils.getAndDeleteFile
 import io.sentry.android.gradle.util.SentryPluginUtils.isMinificationEnabled
 import kotlin.test.assertEquals
+import kotlin.test.assertFalse
+import kotlin.test.assertTrue
+import org.junit.Rule
 import org.junit.Test
+import org.junit.rules.TemporaryFolder
 import proguard.gradle.plugin.android.dsl.ProGuardAndroidExtension
 
 class SentryPluginUtilsTest {
+
+    @get:Rule
+    val tempDir = TemporaryFolder()
 
     @Test
     fun `capitalizes string first letter uppercase`() {
@@ -74,5 +82,16 @@ class SentryPluginUtilsTest {
         val debug = android.applicationVariants.first { it.name == "debug" }
 
         assertEquals(false, isMinificationEnabled(project, debug))
+    }
+
+    @Test
+    fun `getAndDelete deletes the file`() {
+        val (project, _) = createTestAndroidProject()
+        val file = tempDir.newFile("temp-file.txt")
+
+        assertTrue { file.exists() }
+
+        getAndDeleteFile(project.layout.file(project.provider { file }))
+        assertFalse { file.exists() }
     }
 }